Release V25, asyn protocls download & deleting

+++ certoscale/src/main/java/com/certoclav/certoscale/settings/protocol/MenuProtocolActivity.java
@@ -2,17 +2,18 @@ package com.certoclav.certoscale.settings.protocol;
 
 import android.app.Activity;
 import android.app.Dialog;
+import android.app.ProgressDialog;
 import android.content.Intent;
 import android.database.DataSetObserver;
 import android.os.Bundle;
 import android.os.Handler;
 import android.util.Log;
 import android.view.View;
+import android.widget.AbsListView;
 import android.widget.AdapterView;
 import android.widget.Button;
 import android.widget.ListView;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.certoclav.certoscale.R;
 import com.certoclav.certoscale.adapters.ProtocolAdapter;
@@ -34,12 +35,12 @@ import com.google.android.gms.appindexing.AppIndex;
 import com.google.android.gms.common.api.GoogleApiClient;
 
 import java.util.ArrayList;
-import java.util.Calendar;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 
-import es.dmoral.toasty.Toasty;
+import needle.Needle;
+import needle.UiRelatedProgressTask;
 
 import static com.certoclav.certoscale.R.string.protocols;
 
@@ -82,6 +83,9 @@ public class MenuProtocolActivity extends Activity implements ButtonEventListene
         navigationbar.onCreate();
         navigationbar.setButtonEventListener(this);
         navigationbar.getButtonBack().setVisibility(View.VISIBLE);
+        navigationbar.getButtonDelete().setVisibility(Scale.getInstance().getUser().getIsAdmin() ? View.VISIBLE : View.GONE);
+
+
         navigationbar.getTextTitle().setText(getString(protocols).toUpperCase());
         navigationbar.getTextTitle().setVisibility(View.VISIBLE);
         listView = (ListView) findViewById(R.id.menu_main_recipe_list);
@@ -96,6 +100,20 @@ public class MenuProtocolActivity extends Activity implements ButtonEventListene
             }
         });
         listView.setAdapter(adapter);
+
+        listView.setOnScrollListener(new AbsListView.OnScrollListener() {
+            @Override
+            public void onScrollStateChanged(AbsListView absListView, int scrollState) {
+                if (scrollState == 0) {
+                    actionDetected();
+                }
+            }
+
+            @Override
+            public void onScroll(AbsListView absListView, int i, int i1, int i2) {
+
+            }
+        });
 
         listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
@@ -127,59 +145,82 @@ public class MenuProtocolActivity extends Activity implements ButtonEventListene
         adapter.clear();
         Scale.getInstance().setOnDatabaseListener(this);
 
-        List<Protocol> protocolList = getIntent().getBooleanExtra(AppConstants.INTENT_EXTRA_PICK_ON_CLICK, false) ?
-                db.getPengingProtocols() :
-                db.getProtocols();
-        Collections.sort(protocolList, new Comparator<Protocol>() {
-            public int compare(Protocol emp1, Protocol emp2) {
-                return emp2.getDate().compareTo(emp1.getDate()); // To compare string values
+
+        final ProgressDialog progressDialog = new ProgressDialog(MenuProtocolActivity.this);
+        progressDialog.setMessage(getString(R.string.loading));
+        progressDialog.setCancelable(false);
+        progressDialog.show();
+
+
+        Needle.onBackgroundThread().execute(new UiRelatedProgressTask<List<Protocol>, Integer>() {
+            @Override
+            protected List<Protocol> doWork() {
+                List<Protocol> protocolList = getIntent().getBooleanExtra(AppConstants.INTENT_EXTRA_PICK_ON_CLICK, false) ?
+                        db.getPengingProtocols() :
+                        db.getProtocols();
+
+                Collections.sort(protocolList, new Comparator<Protocol>() {
+                    public int compare(Protocol emp1, Protocol emp2) {
+                        return emp2.getDateFormart().compareTo(emp1.getDateFormart()); // To compare string values
+                    }
+                });
+
+                return  protocolList;
+            }
+
+            @Override
+            protected void thenDoUiRelatedWork(List<Protocol> protocolList) {
+                adapter.addAll(protocolList);
+                adapter.notifyDataSetChanged();
+                progressDialog.dismiss();
+            }
+
+            @Override
+            protected void onProgressUpdate(Integer progress) {
             }
         });
 
-        adapter.addAll(protocolList);
-        adapter.notifyDataSetChanged();
-
 //        Intent intent = new Intent(ApplicationController.getContext(), SyncProtocolsService.class);
 //        startService(intent);
-        ftpManager.uploadProtocols(new FTPManager.FTPListener() {
-            @Override
-            public void onConnection(boolean isConnected, final String message) {
-                Log.d("FTP_SERVER", "connection " + isConnected + " " + (message != null ? message : ""));
-                if (!isConnected) {
-                    runOnUiThread(new Runnable() {
-                        @Override
-                        public void run() {
-                            Toasty.error(ApplicationController.getContext(), (message != null && !message.isEmpty())
-                                    ? message : getString(R.string.can_not_connect_to_ftp), Toast.LENGTH_LONG, true).show();
-                        }
-                    });
-
-                }
-            }
-
-            @Override
-            public void onUploaded(Protocol protocol) {
-                if (db != null)
-                    db.updateProtocolCloudId(protocol, "uploaded");
-                Log.d("FTP_SERVER", "updatedall");
-            }
-
-            @Override
-            public void onUploading(boolean isUploaded, final String message) {
-                Log.d("FTP_SERVER", "uploading " + isUploaded + " " + (message != null ? message : ""));
-                if (!isUploaded && lastNotifiedTime + 10 * 1000 < Calendar.getInstance().getTimeInMillis()) {
-                    runOnUiThread(new Runnable() {
-                        @Override
-                        public void run() {
-                            Toasty.error(ApplicationController.getContext(), (message != null && !message.isEmpty())
-                                    ? message : getString(R.string.can_not_uploaded), Toast.LENGTH_SHORT, true).show();
-                        }
-                    });
-
-                    lastNotifiedTime = Calendar.getInstance().getTimeInMillis();
-                }
-            }
-        }, db.getProtocols(), false);
+//        ftpManager.uploadProtocols(new FTPManager.FTPListener() {
+//            @Override
+//            public void onConnection(boolean isConnected, final String message) {
+//                Log.d("FTP_SERVER", "connection " + isConnected + " " + (message != null ? message : ""));
+//                if (!isConnected) {
+//                    runOnUiThread(new Runnable() {
+//                        @Override
+//                        public void run() {
+//                            Toasty.error(ApplicationController.getContext(), (message != null && !message.isEmpty())
+//                                    ? message : getString(R.string.can_not_connect_to_ftp), Toast.LENGTH_LONG, true).show();
+//                        }
+//                    });
+//
+//                }
+//            }
+//
+//            @Override
+//            public void onUploaded(Protocol protocol) {
+//                if (db != null)
+//                    db.updateProtocolCloudId(protocol, "uploaded");
+//                Log.d("FTP_SERVER", "updatedall");
+//            }
+//
+//            @Override
+//            public void onUploading(boolean isUploaded, final String message) {
+//                Log.d("FTP_SERVER", "uploading " + isUploaded + " " + (message != null ? message : ""));
+//                if (!isUploaded && lastNotifiedTime + 10 * 1000 < Calendar.getInstance().getTimeInMillis()) {
+//                    runOnUiThread(new Runnable() {
+//                        @Override
+//                        public void run() {
+//                            Toasty.error(ApplicationController.getContext(), (message != null && !message.isEmpty())
+//                                    ? message : getString(R.string.can_not_uploaded), Toast.LENGTH_SHORT, true).show();
+//                        }
+//                    });
+//
+//                    lastNotifiedTime = Calendar.getInstance().getTimeInMillis();
+//                }
+//            }
+//        }, db.getProtocols(), false);
         actionDetected();
     }
 
@@ -210,6 +251,7 @@ public class MenuProtocolActivity extends Activity implements ButtonEventListene
                     final Dialog dialog = new Dialog(this);
                     dialog.setContentView(R.layout.dialog_yes_no);
                     dialog.setTitle(getString(R.string.delete_all));
+
                     ((TextView) dialog.findViewById(R.id.text)).setText(R.string.do_you_really_want_to_delete_all_protocols);
                     // set the custom dialog components - text, image and button
 
@@ -231,19 +273,52 @@ public class MenuProtocolActivity extends Activity implements ButtonEventListene
                             final DatabaseService db = new DatabaseService(ApplicationController.getContext());
                             actionDetected();
 
+                            final ProgressDialog progressDialog = new ProgressDialog(MenuProtocolActivity.this);
+                            progressDialog.setMessage(getString(R.string.deleting));
+                            progressDialog.setCancelable(false);
+                            progressDialog.show();
+
+
+                            Needle.onBackgroundThread().execute(new UiRelatedProgressTask<Boolean, Integer>() {
+                                @Override
+                                protected Boolean doWork() {
+                                    List<Protocol> protocolList = db.getUploadedProtocols();
+                                    int i = 0;
+                                    int len = protocolList.size();
+                                    if (len == 0)
+                                        return true;
+                                    for (Protocol protocol : protocolList) {
+                                        db.deleteProtocol(protocol);
+                                        i++;
+                                        publishProgress(((i * 100) / len));
+
+                                    }
+                                    return true;
+                                }
+
+                                @Override
+                                protected void thenDoUiRelatedWork(Boolean result) {
+                                    dialog.dismiss();
+                                    onResume();
+                                    progressDialog.dismiss();
+                                }
+
+                                @Override
+                                protected void onProgressUpdate(Integer progress) {
+                                    try {
+                                        actionDetected();
+                                        progressDialog.setMessage(getString(R.string.deleting) + " " + progress + " %");
+                                    } catch (Exception e) {
+
+                                    }
+                                }
+                            });
+
                             //db.getProtocols().removeAll((Collection<Protocol>) db.getProtocols());
-                            List<Protocol> protocolList = db.getProtocols();
 
                                 /*for(int i=0; i<protocolList.size();i++) {
                                     db.deleteProtocol(protocolList.get(i));
                                 }*/
-
-                            for (Protocol protocol : protocolList) {
-                                db.deleteProtocol(protocol);
-                            }
-
-                            dialog.dismiss();
-                            onResume();
 
 
                         }
